+++ go/go.mod
@@ -2,4 +2,9 @@ module github.com/sumanchapai/patro
 
 go 1.21.1
 
-require golang.org/x/exp v0.0.0-20231110203233-9a3e6036ecaa // indirect
+require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/spf13/cobra v1.8.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+	golang.org/x/exp v0.0.0-20231110203233-9a3e6036ecaa // indirect
+)
